refactor: use ScopedCom for icon as well
